Respect editable configuration for standalone editables in headless document
