+++ src/Factory/HeadlessEditableInfoFactory.php
@@ -22,6 +22,7 @@ class HeadlessEditableInfoFactory
             editableId: $info->getId(),
             name: $item['name'],
             type: $item['type'],
+            label: $item['label'] ?? null,
             brickParent: null,
             editableConfiguration: null,
             config: $this->createConfig($info->getDocument(), $item, $editMode, $item['type'] === 'areablock' ? $info->getId() : null),
@@ -38,6 +39,7 @@ class HeadlessEditableInfoFactory
             editableId: $editableId,
             name: $item['name'],
             type: $item['type'],
+            label: $item['label'] ?? null,
             brickParent: null,
             editableConfiguration: $this->createEditableConfiguration($item),
             config: $this->createConfig($document, $item, $editMode, $item['type'] === 'areablock' ? $item['name'] : null),
@@ -86,6 +88,7 @@ class HeadlessEditableInfoFactory
                 editableId: $editableId,
                 name: $columnName,
                 type: 'areablock',
+                label: null,
                 brickParent: $editableId,
                 editableConfiguration: null,
                 config: $this->createConfig($document, $item, $editMode, $editableId),
@@ -126,9 +129,10 @@ class HeadlessEditableInfoFactory
                 editableId: $editableId,
                 name: array_key_exists('name', $childData) ? $childData['name'] : $childName,
                 type: $childData['type'],
+                label: $childData['label'] ?? null,
                 brickParent: $brickParent,
                 editableConfiguration: $editableConfiguration,
-                config: $this->createConfig($document, $childData, $editMode, $editableId),
+                config: $this->createConfig($document, $childData, $editMode, null),
                 params: $parameters,
                 children: [],
                 editMode: $editMode,
